fix: decouple direct image failure details
